requirements.txt
flask==3.0.0
torch==2.2.2
transformers==4.36.2
numpy
opencv-python-headless
pillow
pillow-heif
gunicorn
